feat: add the macro branch for "serialization_check" testing
