+++ src/bn256/fq.rs
@@ -290,6 +290,7 @@ mod test {
     crate::field_testing_suite!(Fq, "serialization");
     crate::field_testing_suite!(Fq, "quadratic_residue");
     crate::field_testing_suite!(Fq, "bits");
+    crate::field_testing_suite!(Fq, "serialization_check");
 
     use crate::ff_ext::Legendre;
     use rand_core::OsRng;
